ai_analytics/services/video_stream_manager.py
"""
Video Stream Manager for the AI Analytics module.

This service manages video streams from cameras and provides frames for processing.
"""

import logging
import time
import asyncio
from typing import Dict, Any, Optional
import cv2
import numpy as np

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class VideoStreamManager:
    """Manager for video streams from cameras."""
    
    def __init__(self, config: Dict[str, Any]):
        """Initialize the video stream manager."""
        self.config = config
        self.streams = {}
        self.frames = {}
        self.last_frame_time = {}
        self.running = False
        self.frame_lock = asyncio.Lock()
    
    async def initialize(self):
        """Initialize video streams."""
        try:
            logger.info("Initializing video streams")
            
            # Start background task to fetch frames
            self.running = True
            asyncio.create_task(self._fetch_frames_task())
            
            logger.info("Video streams initialized")
        except Exception as e:
            logger.error(f"Error initializing video streams: {str(e)}")
            raise
    
    async def shutdown(self):
        """Shut down video streams."""
        try:
            logger.info("Shutting down video streams")
            
            # Stop background task
            self.running = False
            
            # Release all streams
            for camera_id, stream in self.streams.items():
                if stream:
                    # In a real implementation, you would release the stream
                    # For example: stream.release()
                    logger.info(f"Released stream for camera {camera_id}")
            
            # Clear streams and frames
            self.streams.clear()
            self.frames.clear()
            self.last_frame_time.clear()
            
            logger.info("Video streams shut down")
        except Exception as e:
            logger.error(f"Error shutting down video streams: {str(e)}")
            raise
    
    async def _fetch_frames_task(self):
        """Background task to fetch frames from all streams."""
        try:
            while self.running:
                # Fetch frames from all streams
                for camera_id in self.get_camera_ids():
                    try:
                        await self._fetch_frame(camera_id)
                    except Exception as e:
                        logger.error(f"Error fetching frame from camera {camera_id}: {str(e)}")
                
                # Sleep to avoid high CPU usage
                await asyncio.sleep(0.01)
        except Exception as e:
            logger.error(f"Error in fetch frames task: {str(e)}")
            self.running = False
    
    async def _fetch_frame(self, camera_id: str):
        """Fetch a frame from a camera stream."""
        # This is a placeholder for actual frame fetching
        # In a real implementation, you would fetch a frame from the camera stream
        
        # Check if stream exists
        if camera_id not in self.streams:
            # Initialize stream
            await self._initialize_stream(camera_id)
        
        # Get stream
        stream = self.streams.get(camera_id)
        
        if not stream:
            return
        
        # In a real implementation, you would fetch a frame from the stream
        # For example: ret, frame = stream.read()
        
        # For now, we'll just simulate a frame
        frame = np.zeros((480, 640, 3), dtype=np.uint8)
        
        # Add timestamp to frame
        timestamp = time.time()
        cv2.putText(
            frame,
            f"Camera {camera_id} - {time.strftime('%Y-%m-%d %H:%M:%S')}",
            (10, 30),
            cv2.FONT_HERSHEY_SIMPLEX,
            0.7,
            (255, 255, 255),
            2
        )
        
        # Store frame
        async with self.frame_lock:
            self.frames[camera_id] = frame
            self.last_frame_time[camera_id] = timestamp
    
    async def _initialize_stream(self, camera_id: str):
        """Initialize a camera stream."""
        try:
            logger.info(f"Initializing stream for camera {camera_id}")
            
            # This is a placeholder for actual stream initialization
            # In a real implementation, you would initialize a stream from the camera
            # For example: stream = cv2.VideoCapture(camera_url)
            
            # For now, we'll just simulate a stream
            stream = {"camera_id": camera_id, "initialized": True}
            
            # Store stream
            self.streams[camera_id] = stream
            
            logger.info(f"Stream initialized for camera {camera_id}")
        except Exception as e:
            logger.error(f"Error initializing stream for camera {camera_id}: {str(e)}")
            self.streams[camera_id] = None
    
    def get_camera_ids(self) -> list:
        """Get all camera IDs."""
        # This is a placeholder for actual camera ID retrieval
        # In a real implementation, you would retrieve camera IDs from configuration or discovery
        
        # Example camera IDs
        return ["camera1", "camera2", "camera3", "camera4"]
    
    async def get_frame(self, camera_id: str) -> Optional[np.ndarray]:
        """Get the latest frame from a camera."""
        # Check if frame exists
        async with self.frame_lock:
            frame = self.frames.get(camera_id)
            
            if frame is None:
                # Try to fetch frame
                await self._fetch_frame(camera_id)
                frame = self.frames.get(camera_id)
            
            # Return copy of frame to avoid modification
            return frame.copy() if frame is not None else None
    
    async def get_frame_timestamp(self, camera_id: str) -> Optional[float]:
        """Get the timestamp of the latest frame from a camera."""
        async with self.frame_lock:
            return self.last_frame_time.get(camera_id)
