ai_analytics/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
opencv-python>=4.7.0
PyYAML>=6.0
aiofiles>=23.1.0
httpx>=0.24.0
redis>=4.5.4
ultralytics>=8.0.0
face-recognition>=1.3.0
scikit-learn>=1.2.2
pandas>=2.0.0
matplotlib>=3.7.1
prophet>=1.1.4
xgboost>=1.7.5
tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.1
onnxruntime>=1.14.1
pytesseract>=0.3.10
Pillow>=9.5.0
